Rename SMW_setup.php to setupStore.php
- The script has several issues but are not addressed in this PR
- Deprecate SMW_setup.php
- Add SetupStoreMaintenanceRegressionTest
